+++ Realm/RLMSyncManager_Private.h
@@ -19,6 +19,7 @@
 #import <Realm/RLMSyncManager.h>
 
 #import "RLMSyncUtil_Private.h"
+#import "RLMNetworkClient.h"
 
 @class RLMSyncUser, RLMSyncConfiguration;
 
@@ -43,6 +44,8 @@ NS_ASSUME_NONNULL_BEGIN
 - (NSArray<RLMSyncUser *> *)_allUsers;
 
 + (void)resetForTesting;
+
+- (RLMNetworkRequestOptions *)networkRequestOptions;
 
 NS_ASSUME_NONNULL_END